Move routes to own directory
